# watchdog/openapi.yaml
openapi: 3.1.0
info:
  title: Watchdog with Your Infrastructure
  description: Unified API for managing Tempo, Loki, AlertManager, Grafana, and Resolver
  version: 0.0.1
paths:
...
  /api/auth/me:
    get:
      tags:
      - authentication
      summary: Get Current User Info
      operationId: get_current_user_info_api_auth_me_get
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
      security:
      - HTTPBearer: []
    put:
      tags:
      - authentication
      summary: Update Current User Info
      operationId: update_current_user_info_api_auth_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
      security:
      - HTTPBearer: []
  /api/auth/users:
    get:
      tags:
      - authentication
      summary: List Users
      operationId: list_users_api_auth_users_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000000
          minimum: 0
          default: 0
          title: Offset
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
    post:
      tags:
      - authentication
      summary: Create User
      operationId: create_user_api_auth_users_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
  /api/auth/users/{user_id}:
    put:
      tags:
      - authentication
      summary: Update User
      operationId: update_user_api_auth_users__user_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
    delete:
      tags:
      - authentication
      summary: Delete User
      operationId: delete_user_api_auth_users__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: User Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /api/auth/users/{user_id}/password:
    put:
      tags:
      - authentication
      summary: Update User Password
      operationId: update_user_password_api_auth_users__user_id__password_put
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPasswordUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /api/auth/users/{user_id}/password/reset-temp:
    post:
      tags:
      - authentication
      summary: Reset User Password Temp
      operationId: reset_user_password_temp_api_auth_users__user_id__password_reset_temp_post
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: User Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TempPasswordResetResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /api/auth/users/{user_id}/permissions:
    put:
      tags:
      - authentication
      summary: Update User Permissions
      operationId: update_user_permissions_api_auth_users__user_id__permissions_put
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
              title: Permission Names
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /api/auth/permissions:
    get:
      tags:
      - authentication
      summary: List All Permissions
      operationId: list_all_permissions_api_auth_permissions_get
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                items:
                  additionalProperties:
                    $ref: '#/components/schemas/JSONValue'
                  type: object
                type: array
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
      security:
      - HTTPBearer: []
  /api/auth/role-defaults:
    get:
      tags:
      - authentication
      summary: List Role Defaults
      operationId: list_role_defaults_api_auth_role_defaults_get
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    type: string
                  type: array
                type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    ApiKey:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          pattern: ^[^\x00-\x1F]+$
          title: Name
        id:
          type: string
          title: Id
        key:
          type: string
          title: Key
        otlp_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Otlp Token
          description: Secure OTLP ingest token for gateway authentication
        owner_user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner User Id
        owner_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Username
        is_shared:
          type: boolean
          title: Is Shared
          default: false
        can_use:
          type: boolean
          title: Can Use
          default: true
        shared_with:
          items:
            $ref: '#/components/schemas/ApiKeyShareUser'
          type: array
          title: Shared With
        is_default:
          type: boolean
          title: Is Default
          default: false
        is_enabled:
          type: boolean
          title: Is Enabled
          default: true
        is_hidden:
          type: boolean
          title: Is Hidden
          default: false
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
      type: object
      required:
      - name
      - id
      - key
      - created_at
      title: ApiKey
...
    ApiKeyShareUser:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        can_use:
          type: boolean
          title: Can Use
          default: true
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - user_id
      - created_at
      title: ApiKeyShareUser
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Permission:
      type: string
      enum:
      - read:audit_logs
      - read:alerts
      - create:alerts
      - update:alerts
      - write:alerts
      - delete:alerts
      - read:silences
      - create:silences
      - update:silences
      - delete:silences
      - read:rules
      - create:rules
      - update:rules
      - delete:rules
      - test:rules
      - read:metrics
      - read:channels
      - create:channels
      - update:channels
      - write:channels
      - delete:channels
      - test:channels
      - read:incidents
      - update:incidents
      - read:logs
      - read:traces
      - read:rca
      - create:rca
      - delete:rca
      - read:dashboards
      - create:dashboards
      - update:dashboards
      - write:dashboards
      - delete:dashboards
      - read:datasources
      - create:datasources
      - update:datasources
      - delete:datasources
      - query:datasources
      - read:folders
      - create:folders
      - delete:folders
      - read:agents
      - read:api_keys
      - create:api_keys
      - update:api_keys
      - delete:api_keys
      - create:users
      - update:users
      - delete:users
      - update:user_permissions
      - manage:users
      - read:users
      - create:groups
      - update:groups
      - delete:groups
      - update:group_permissions
      - update:group_members
      - manage:groups
      - read:groups
      - manage:tenants
      title: Permission
...
    Role:
      type: string
      enum:
      - provisioning
      - admin
      - user
      - viewer
      title: Role
...
    TempPasswordResetResponse:
      properties:
        temporary_password:
          type: string
          title: Temporary Password
        email_sent:
          type: boolean
          title: Email Sent
        message:
          type: string
          title: Message
      type: object
      required:
      - temporary_password
      - email_sent
      - message
      title: TempPasswordResetResponse
...
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 3
          title: Username
        email:
          type: string
          format: email
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        org_id:
          type: string
          maxLength: 100
          title: Org Id
          description: Organization ID for multi-tenant observability
          default: default
        role:
          $ref: '#/components/schemas/Role'
          default: user
        group_ids:
          items:
            type: string
          type: array
          title: Group Ids
        is_active:
          type: boolean
          title: Is Active
          default: true
        password:
          anyOf:
          - type: string
            minLength: 8
          - type: 'null'
          title: Password
        must_setup_mfa:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Must Setup Mfa
      additionalProperties: false
      type: object
      required:
      - username
      - email
      title: UserCreate
    UserPasswordUpdate:
      properties:
        current_password:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Password
        new_password:
          type: string
          minLength: 8
          title: New Password
      type: object
      required:
      - new_password
      title: UserPasswordUpdate
    UserResponse:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        role:
          $ref: '#/components/schemas/Role'
        group_ids:
          items:
            type: string
          type: array
          title: Group Ids
        is_active:
          type: boolean
          title: Is Active
        org_id:
          type: string
          title: Org Id
        tenant_id:
          type: string
          title: Tenant Id
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        last_login:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Login
        permissions:
          items:
            $ref: '#/components/schemas/Permission'
          type: array
          title: Permissions
        direct_permissions:
          items:
            type: string
          type: array
          title: Direct Permissions
        needs_password_change:
          type: boolean
          title: Needs Password Change
          default: false
        api_keys:
          items:
            $ref: '#/components/schemas/ApiKey'
          type: array
          title: Api Keys
        mfa_enabled:
          type: boolean
          title: Mfa Enabled
          default: false
        must_setup_mfa:
          type: boolean
          title: Must Setup Mfa
          default: false
        auth_provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Auth Provider
          default: local
      type: object
      required:
      - id
      - username
      - email
      - full_name
      - role
      - group_ids
      - is_active
      - org_id
      - tenant_id
      - created_at
      - last_login
      - permissions
      title: UserResponse
    UserUpdate:
      properties:
        username:
          anyOf:
          - type: string
            maxLength: 50
            minLength: 3
          - type: 'null'
          title: Username
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        org_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Org Id
        role:
          anyOf:
          - $ref: '#/components/schemas/Role'
          - type: 'null'
        group_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Group Ids
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        must_setup_mfa:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Must Setup Mfa
      additionalProperties: false
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    JSONValue:
      anyOf: *id002
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    WatchdogCookieAuth:
      type: apiKey
      in: cookie
      name: watchdog_token
      description: Authenticate using the watchdog_token cookie (same token as Bearer).
...